Tests/LibWeb/Text/expected/navigate-renderable-not-downloaded.txt

download requested: false
iframe rendered the response: true
